refactor: 공용 컴포넌트 타입 파일 분리 및 적용
