+++ src/components/composables/Select.ts
@@ -1,4 +1,4 @@
-import { OptionElementPropsType, SelectElementPropsType, SelectPropsType } from '../../types';
+import { OptionElementPropsType, SelectElementPropsType, SelectPropsType } from '../../types/components';
 
 const createSelectElement = ({ id, name, className }: SelectElementPropsType) =>
   Object.assign(document.createElement('select'), {
