+++ src/uiUtils/generateInputComponent.ts
@@ -1,0 +1,3 @@
+const generateInputComponent = (inputOptionData) => {};
+
+export default generateInputComponent;
